+++ soc_log_anonymizer/__init__.py
@@ -11,7 +11,7 @@ from .config import AnonymizerConfig
 from .io_utils import read_file_auto_encoding
 from .result import AnonymizeResult
 
-__version__ = "2.4.0"
+__version__ = "2.5.0"
 
 __all__ = [
     "AnonymizeResult",